requirements.txt
fdroidserver==2.4.3
pyyaml==6.0.2
